test: consolidate page tests using editPagesCommonTests utility
